+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Next
 
 - recording: mask swiftui picker if masking enabled ([#184](https://github.com/PostHog/posthog-ios/pull/184))
+- chore: add is identified property ([#186](https://github.com/PostHog/posthog-ios/pull/186))
 
 ## 3.9.1 - 2024-09-06